requirements.txt
langchain
streamlit
python-dotenv
openai
langchain-community